% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_monthly_gridded}
\alias{download_monthly_gridded}
\title{Download gridded data based on indices}
\usage{
download_monthly_gridded(index, dir_path)
}
\arguments{
\item{index}{url file paths for download}

\item{dir_path}{folder to download files to}
}
\value{
file paths
}
\description{
Download gridded data based on indices
}
