% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_time_series}
\alias{get_time_series}
\title{Time series extraction from gridded data}
\usage{
get_time_series(file_paths, lat, lon, project = TRUE)
}
\arguments{
\item{file_paths}{character, file names for read-in; these are returned from \code{\link{download_monthly_gridded}}.}

\item{lat}{numeric, latitude of site}

\item{lon}{numeric, longitude of site}
}
\value{
data.frame with time series for input files and site
}
\description{
Should work with multiple sites, but not tested (i.e. \code{lat = c(x,y,z)})).
}
