% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{index_year_mon}
\alias{index_year_mon}
\title{Get time info from file paths}
\usage{
index_year_mon(index)
}
\arguments{
\item{index}{character, url paths or file names for rdwd download / read-in}
}
\value{
Matrix with time info from input files/urls
}
\description{
Get time info from file paths
}
