% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_monthly_gridded_indices}
\alias{get_monthly_gridded_indices}
\title{Get urls for monthly gridded min and max temp}
\usage{
get_monthly_gridded_indices(gridIndex)
}
\arguments{
\item{gridIndex}{character, set index from \code{data("gridIndex", package = "rdwd")}}
}
\value{
urls for file download; function is hard coded for air temp min/max and precip
}
\description{
Get urls for monthly gridded min and max temp
}
